QuetSDT_Facebook_Auto/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
selenium==4.16.0
webdriver-manager==4.0.1
pandas==2.1.4
python-dotenv==1.0.0